+++ frontend/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import Backdrop from './components/Backdrop/Backdrop.js'
-import Toolbar from './components/Toolbar/Toolbar.js'
+import Toolbar from './containers/Toolbar/Toolbar.js'
 
 function App() {
   return (
